requirements.txt
streamlit
requests
pyyaml
julep